requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
azure-core==1.38.0
azure-identity==1.25.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
dnspython==2.7.0
exceptiongroup==1.3.1
fastapi==0.124.4
Flask==3.1.2
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
pillow==11.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.5
pyodbc==5.3.0
python-dotenv==1.2.1
reportlab==4.4.6
requests==2.32.5
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.38.0
Werkzeug==3.1.5
zipp==3.23.0
